backend/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14
mysqlclient>=2.1
django-cors-headers>=3.14
python-decouple>=3.8